Split off basic spline generation into separate class Split base_trajectory matlab printing into separate file
