+++ zebROS_ws/src/base_trajectory/include/base_trajectory/matlab_printer.h
@@ -1,0 +1,6 @@
+
+#include "base_trajectory/GenerateSpline.h"
+
+void writeMatlabSplines(const base_trajectory::GenerateSpline::Response &msg);
+void writeMatlabPath(const std::vector<geometry_msgs::PoseStamped> &poses);
+
