+++ zebROS_ws/src/spline_util/package.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>base_trajectory</name>
+  <name>spline_util</name>
   <version>0.0.0</version>
-  <description>The base_trajectory package</description>
+  <description>The spline_util package</description>
 
   <!-- One maintainer tag required, multiple allowed, one person per tag -->
   <!-- Example:  -->
   <!-- <maintainer email="jane.doe@example.com">Jane Doe</maintainer> -->
-  <maintainer email="programmers@team900.org">kjaget</maintainer>
+  <maintainer email="kjaget@todo.todo">kjaget</maintainer>
 
 
   <!-- One license tag required, multiple allowed, one license per tag -->
@@ -19,7 +19,7 @@
   <!-- Url tags are optional, but multiple are allowed, one per tag -->
   <!-- Optional attribute type can be: website, bugtracker, or repository -->
   <!-- Example: -->
-  <!-- <url type="website">http://wiki.ros.org/base_trajectory</url> -->
+  <!-- <url type="website">http://wiki.ros.org/spline_util</url> -->
 
 
   <!-- Author tags are optional, multiple are allowed, one per tag -->
@@ -49,15 +49,17 @@
   <!-- Use doc_depend for packages you need only for building documentation: -->
   <!--   <doc_depend>doxygen</doc_depend> -->
   <buildtool_depend>catkin</buildtool_depend>
-  <depend>ddynamic_reconfigure</depend>
-  <depend>joint_trajectory_controller</depend>
-  <depend>roscpp</depend>
-  <depend>spline_util</depend>
-  <depend>angles</depend>
-  <depend>nav_msgs</depend>
-  <build_depend>message_generation</build_depend>
-  <exec_depend>message_runtime</exec_depend>
+  <build_depend>joint_trajectory_controller</build_depend>
+  <build_depend>roscpp</build_depend>
+  <build_export_depend>joint_trajectory_controller</build_export_depend>
+  <build_export_depend>roscpp</build_export_depend>
+  <exec_depend>joint_trajectory_controller</exec_depend>
+  <exec_depend>roscpp</exec_depend>
+
+
   <!-- The export tag contains other, unspecified, tags -->
   <export>
+    <!-- Other tools can request additional information be placed here -->
+
   </export>
 </package>
